This Is My Portfolio Website
Hello! I'm Rahul Paramanik, and this website is my personal portfolio. As a 2nd-year BCA student, I'm committed to honing my skills in web development and programming. This site, built from the ground up with HTML, CSS, and JavaScript, is a testament to that dedication. Within, you'll find a collection of my projects, each demonstrating my commitment to clean code, effective design, and cutting-edge web technologies. I'm eager to apply my knowledge and skills to create innovative and effective solutions.